Particle: State to class enum

+++ src/particle/src/engine/4C_particle_engine_container.hpp
@@ -160,7 +160,7 @@ namespace Particle
       FOUR_C_ASSERT(storedstates_.contains(state), "particle state '{}' not stored in container!",
           enum_to_state_name(state));
 
-      return statedim_[state];
+      return statedim_[static_cast<int>(state)];
     };
 
     //! \name access global id and particle states
@@ -289,7 +289,8 @@ namespace Particle
 
       double* state_ptr = get_ptr_to_state_writable(state, 0);
 
-      for (int i = 0; i < (particlestored_ * statedim_[state]); ++i) state_ptr[i] *= fac;
+      for (int i = 0; i < (particlestored_ * statedim_[static_cast<int>(state)]); ++i)
+        state_ptr[i] *= fac;
     };
 
     /*!
@@ -314,14 +315,15 @@ namespace Particle
       FOUR_C_ASSERT(storedstates_.contains(stateB), "particle state '{}' not stored in container!",
           enum_to_state_name(stateB));
 
-      FOUR_C_ASSERT(statedim_[stateA] == statedim_[stateB], "dimensions of states do not match!");
+      FOUR_C_ASSERT(statedim_[static_cast<int>(stateA)] == statedim_[static_cast<int>(stateB)],
+          "dimensions of states do not match!");
 
       if (particlestored_ <= 0) return;
 
       const double* state_b_ptr = get_ptr_to_state(stateB, 0);
       double* state_a_ptr = get_ptr_to_state_writable(stateA, 0);
 
-      for (int i = 0; i < (particlestored_ * statedim_[stateA]); ++i)
+      for (int i = 0; i < (particlestored_ * statedim_[static_cast<int>(stateA)]); ++i)
         state_a_ptr[i] = facA * state_a_ptr[i] + facB * state_b_ptr[i];
     };
 
@@ -337,16 +339,16 @@ namespace Particle
       FOUR_C_ASSERT(storedstates_.contains(state), "particle state '{}' not stored in container!",
           enum_to_state_name(state));
 
-      FOUR_C_ASSERT(
-          statedim_[state] == static_cast<int>(val.size()), "dimensions of states do not match!");
+      FOUR_C_ASSERT(statedim_[static_cast<int>(state)] == static_cast<int>(val.size()),
+          "dimensions of states do not match!");
 
       if (particlestored_ <= 0) return;
 
       double* state_ptr = get_ptr_to_state_writable(state, 0);
 
       for (int i = 0; i < particlestored_; ++i)
-        for (int dim = 0; dim < statedim_[state]; ++dim)
-          state_ptr[i * statedim_[state] + dim] = val[dim];
+        for (int dim = 0; dim < statedim_[static_cast<int>(state)]; ++dim)
+          state_ptr[i * statedim_[static_cast<int>(state)] + dim] = val[dim];
     };
 
     /*!
@@ -364,7 +366,8 @@ namespace Particle
 
       double* state_ptr = get_ptr_to_state_writable(state, 0);
 
-      for (int i = 0; i < (particlestored_ * statedim_[state]); ++i) state_ptr[i] = 0.0;
+      for (int i = 0; i < (particlestored_ * statedim_[static_cast<int>(state)]); ++i)
+        state_ptr[i] = 0.0;
     };
 
     //! @}
